sanity commit

+++ app/src/main/java/com/abasscodes/myapplication/model/api/Rates.java
@@ -50,7 +50,7 @@ public class Rates implements Parcelable{
     private Double hUF;
     @SerializedName("IDR")
     @Expose
-    private Integer iDR;
+    private Double iDR;
     @SerializedName("ILS")
     @Expose
     private Double iLS;
@@ -292,14 +292,14 @@ public class Rates implements Parcelable{
     /**
      * @return The iDR
      */
-    public Integer getIDR() {
+    public Double getIDR() {
         return iDR;
     }
 
     /**
      * @param iDR The IDR
      */
-    public void setIDR(Integer iDR) {
+    public void setIDR(Double iDR) {
         this.iDR = iDR;
     }
